+++ allennlp/modules/seq2vec_encoders/cnn_encoder.py
@@ -84,7 +84,7 @@ class CnnEncoder(Seq2VecEncoder):
 
     def forward(self, tokens: torch.Tensor, mask: torch.Tensor):  # pylint: disable=arguments-differ
         if mask is not None:
-            tokens = tokens * mask.unsqueeze(-1).float()
+            tokens = tokens * mask.unsqueeze(-1).type_as(tokens)
 
         # Our input is expected to have shape `(batch_size, num_tokens, embedding_dim)`.  The
         # convolution layers expect input of shape `(batch_size, in_channels, sequence_length)`,
