+++ allennlp/modules/span_extractors/self_attentive_span_extractor.py
@@ -75,13 +75,13 @@ class SelfAttentiveSpanExtractor(SpanExtractor):
         # We're using <= here (and for the mask below) because the span ends are
         # inclusive, so we want to include indices which are equal to span_widths rather
         # than using it as a non-inclusive upper bound.
-        span_mask = (max_span_range_indices <= span_widths).float()
+        span_mask = (max_span_range_indices <= span_widths)
         raw_span_indices = span_ends - max_span_range_indices
         # We also don't want to include span indices which are less than zero,
         # which happens because some spans near the beginning of the sequence
         # have an end index < max_batch_span_width, so we add this to the mask here.
-        span_mask = span_mask * (raw_span_indices >= 0).float()
-        span_indices = torch.nn.functional.relu(raw_span_indices.float()).long()
+        span_mask = span_mask * (raw_span_indices >= 0)
+        span_indices = torch.nn.functional.relu(raw_span_indices.type_as(sequence_tensor)).long()
 
         # Shape: (batch_size * num_spans * max_batch_span_width)
         flat_span_indices = util.flatten_and_batch_shift_indices(span_indices, sequence_tensor.size(1))
@@ -105,6 +105,6 @@ class SelfAttentiveSpanExtractor(SpanExtractor):
             # Above we were masking the widths of spans with respect to the max
             # span width in the batch. Here we are masking the spans which were
             # originally passed in as padding.
-            return attended_text_embeddings * span_indices_mask.unsqueeze(-1).float()
+            return attended_text_embeddings * span_indices_mask.unsqueeze(-1).type_as(attended_text_embeddings)
 
         return attended_text_embeddings
